generalizing badge in AppHeaderLink
