+++ packages/gamut-labs/src/AppHeader/AppHeaderElements/AppHeaderLinkSections/index.tsx
@@ -1,5 +1,3 @@
-import { Badge } from '@codecademy/gamut';
-import styled from '@emotion/styled';
 import React from 'react';
 
 import { AppHeaderLinkMobile } from '../../../AppHeaderMobile/AppHeaderLinkMobile';
@@ -13,32 +11,6 @@ export type AppHeaderLinkSectionsProps = {
   action: AppHeaderClickHandler;
   item: AppHeaderDropdownItem;
 };
-
-export const StyledBadge = styled(Badge)`
-  font-size: 10px;
-  padding-top: 1px;
-  padding-bottom: 10px;
-  padding-left: 8px;
-  padding-right: 8px;
-  width: 36px;
-  border-radius: 16px;
-`;
-
-const renderBadge = (shouldRender?: boolean) =>
-  shouldRender ? (
-    <StyledBadge
-      round
-      variant="blue"
-      fontFamily="accent"
-      ml={8}
-      pl={8}
-      pr={8}
-      height={16}
-      alignSelf="center"
-    >
-      New
-    </StyledBadge>
-  ) : undefined;
 
 export const AppHeaderLinkSections: React.FC<AppHeaderLinkSectionsProps> = ({
   action,
@@ -55,7 +27,7 @@ export const AppHeaderLinkSections: React.FC<AppHeaderLinkSectionsProps> = ({
                   item={link}
                   key={link.id}
                   topSeparator={sectionIndex !== 0 && linkIndex === 0}
-                  badge={renderBadge(link.hasBadge)}
+                  badge={link.badge}
                 />
               );
             });
@@ -66,7 +38,7 @@ export const AppHeaderLinkSections: React.FC<AppHeaderLinkSectionsProps> = ({
                 action={action}
                 item={link}
                 key={link.id}
-                badge={renderBadge(link.hasBadge)}
+                badge={link.badge}
               />
             );
           })}
