+++ packages/gamut-labs/src/AppHeader/AppHeaderElements/types.tsx
@@ -1,3 +1,4 @@
+import { BadgeProps } from '@codecademy/gamut';
 import { GamutIconProps } from '@codecademy/gamut-icons';
 import { ReactNode } from 'react';
 
@@ -34,7 +35,7 @@ export type AppHeaderLinkItem = AppHeaderBaseHrefItem<'link'> & {
   newTab?: boolean;
   text: string;
   topSeparator?: boolean;
-  hasBadge?: boolean;
+  badge?: React.ReactElement<BadgeProps>;
 };
 
 export type AppHeaderTextButtonItem = AppHeaderBaseHrefItem<'text-button'> & {
